platform/metal: trigger bootkube/tectonic via path unit
This removes the need for the last ssh command to the bootkubing
master, which was triggering/running the bootkube/tectonic service
manually in a synchronous way.
This allows for the bootkubing process to be carried-on across
reboots and other failures. The new bootkube trigger is based
on the existence of the shell scripts in /opt.