modules/ignition/kubelet: set /var/lib/cni/bin as --cni-bin-dir
Currently we conditionally set --cni-bin-dir=/var/lib/cni/bin in the
kubelet if network policy is enabled.

In fact, flannel copies binaries to this directory too, but since in this case the
flag is unset, the kubelet defaults to /opt/cni/bin. CNI binaries work
because they are included in the hyperkube image but the
aforementioned copied flannel binaries are effectively unused.

This fixes it.

Fixes INST-532